refactor: 拆分 main.py 逻辑，新增 Scanner 模块
