+++ Scanner/__init__.py
@@ -1,0 +1,4 @@
+from Scanner.scan import Scanner
+from Scanner.init import load_initial_urls, create_duplicate_checker
+
+__all__ = ['Scanner', 'load_initial_urls', 'create_duplicate_checker']
